Re-route categories
