+++ templates/admin/admin/index.html.twig
@@ -9,7 +9,7 @@
                 <h2>Admin Dashboard</h2>
                 <ul>
                     <li><a href="{{ path('app_admin_user_index') }}">Users</a></li>
-                    <li><a href="{{ path('app_admin_category') }}">Categories</a></li>
+                    <li><a href="{{ path('app_admin_category_index') }}">Categories</a></li>
                     <li><a href="{{ path('app_admin_post_index') }}">Posts</a></li>
                     <li><a href="{{ path('app_admin_comment_index') }}">Comments</a></li>
                 </ul>
