+++ templates/admin/category/_delete_form.html.twig
@@ -1,0 +1,4 @@
+<form method="post" action="{{ path('app_admin_category_delete', {'id': category.id}) }}" onsubmit="return confirm('Are you sure you want to delete this item?');">
+    <input type="hidden" name="_token" value="{{ csrf_token('delete' ~ category.id) }}">
+    <button class="btn">Delete</button>
+</form>
