+++ templates/admin/category/edit.html.twig
@@ -5,9 +5,9 @@
 {% block body %}
     <h1>Edit Category</h1>
 
-    {{ include('category/_form.html.twig', {'button_label': 'Update'}) }}
+    {{ include('admin/category/_form.html.twig', {'button_label': 'Update'}) }}
 
     <a href="{{ path('app_home') }}">back to home</a>
 
-    {{ include('category/_delete_form.html.twig') }}
+    {{ include('admin/category/_delete_form.html.twig') }}
 {% endblock %}
