+++ templates/admin/category/new.html.twig
@@ -5,7 +5,7 @@
 {% block body %}
     <h1>Create new Category</h1>
 
-    {{ include('category/_form.html.twig') }}
+    {{ include('admin/category/_form.html.twig') }}
 
-    <a href="{{ path('app_category_index') }}">back to list</a>
+    <a href="{{ path('app_admin_category_index') }}">back to list</a>
 {% endblock %}
