--- templates/category/_delete_form.html.twig
@@ -1,4 +1,0 @@
-<form method="post" action="{{ path('app_category_delete', {'id': category.id}) }}" onsubmit="return confirm('Are you sure you want to delete this item?');">
-    <input type="hidden" name="_token" value="{{ csrf_token('delete' ~ category.id) }}">
-    <button class="btn">Delete</button>
-</form>
